Stdlib: round() precision and PHP_ROUND_* modes — VM + AOT (#3522)
Implement php-src _php_math_round in VmRound/JitRound with __compiler_round
native helper, fix php-types return type (float not int), and add compliance
plus AOT fixtures. MCJIT execute remains deferred like other __compiler_* math.

+++ prelinked/bootstrap-vendor/sources/ircmaxell/php-types/lib/PHPTypes/InternalArgInfo.php
@@ -6422,7 +6422,7 @@ class InternalArgInfo
         'rewind' => ['bool', 'fp' => ''],
         'rewinddir' => ['', 'dir_handle=' => ''],
         'rmdir' => ['bool', 'dirname' => 'string', 'context=' => ''],
-        'round' => ['int', 'number' => 'float', 'precision=' => 'int', 'mode=' => 'int'],
+        'round' => ['float', 'number' => 'float', 'precision=' => 'int', 'mode=' => 'int'],
         'rpm_close' => ['bool', 'rpmr' => ''],
         'rpm_get_tag' => ['', 'rpmr' => '', 'tagnum' => 'int'],
         'rpm_is_valid' => ['bool', 'filename' => 'string'],
